put explict cast to float
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@21807 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: ed20475cfb65f8c469c675d779589b3f212c3c4c

+++ sandbox/functional_m3n/src/regressors/regression_tree_wrapper.cpp
@@ -319,7 +319,7 @@ int RegressionTreeWrapper::train()
       NULL, // sample_idx
       var_type, //var_type
       NULL, // missing mask
-      CvDTreeParams(static_cast<int> (rtree_params_.max_tree_depth_factor * stacked_feature_dim_), // max depth
+      CvDTreeParams(static_cast<int> (rtree_params_.max_tree_depth_factor * static_cast<float>(stacked_feature_dim_)), // max depth
           rtree_params_.min_sample_count, // min sample count
           rtree_params_.regression_accuracy, // regression accuracy
           false, // do NOT compute surrogate split (no missing data)
